// ecis-nodejs-zgwx-feat_240603/service/account-zgxw/ecissdk/model/data.ts
export interface YundocCreateFolderResp {
  result: string
  data: any
  msg?: string
}

export interface YundocCreateFilesUploadResp {
  result: string
  data: any
  msg?: string
}
export interface YundocCreateFileResp {
  result: string
  data: any
  msg?: string
}
export interface YundocUpdateFileResp {
  result: string
  data: { uploadinfo: any }
  msg?: string
}

export interface YundocDownloadFileResp {
  result: string
  data: any
  msg?: string
}

export interface YundocBatchDownloadFileResp {
  result: string
  data: any
  msg?: string
}
export interface YundocListFilesResp {
  result: string
  data: any
  msg?: string
}
export interface YundocGetFilePermissionsResp {
  result: string
  data: any
  msg?: string
}

export interface YundocDeleteFilePermissionsResp {
  result: string
  data: any
  msg?: string
}
export interface Headers {
  [key: string]: any
}
export interface ListFilesParams {
  query: {
    count: number
    filter: string
    groupid: number
    ignore: Array<string>
    include: Array<string>
    include_exts: string
    linkgroup: boolean
    offset: number
    order: string
    orderby: string
    parentid: number
    review_pic_thumbnail: boolean
  }
  headers: Headers
}

export interface CreateFolderParams {
  body: {
    groupid: number
    must_create: boolean
    name: string
    parent_path: Array<string>
    parentid: number
    retrieve_existent: boolean
    csrfmiddlewaretoken: string
  }
  headers: Headers
}

export interface GetFilePermissionsParams{
  query: {
    groupId: number
    fileId: number
    next: string
  }
  headers: Headers
}

export interface DeleteFilePermissionParams{
  query: {
    groupId: number
    fileId: number
    subject_id: number
    subject_type: string
  }
  headers: Headers
}

export interface CreateFilesUploadParams{
  body: {
    checkname: boolean
    client_stores: string
    contenttype: string
    groupid: number
    name: string
    parentid: number
    req_by_internal: boolean
    size: number
    startswithfilename: string
    successactionstatus: string
    trytime: number,
    sha1: string,
    csrfmiddlewaretoken: string
  }
  req: {
    method: string
  }
  headers: Headers
}

export interface CreateFileParams{
  body: {
    etag: string
    fileid: number
    groupid: number
    key: string
    mac:string
    must_create: boolean
    name: string
    parent_path: Array<string>
    parentid: number
    secure_guid: string
    sha1: string
    size: number
    store: string
    unlimited_size: boolean
    csrfmiddlewaretoken: string
  }
  req: {
    method: string
  }
  headers: Headers
}
export interface UpdateFileParams {
  body: {
    client_stores: string
    fileid: number
    req_by_internal: boolean
    s3sha256: string
    sha1: string
    size: number
    trytime: number
    csrfmiddlewaretoken: string
  }
  headers: Headers
}

export interface DownloadFileParams {
  query: {
    fileid: number
    groupid: number
    isblocks: boolean
    processon_type: string
    store: string
  }
  headers: Headers
}

export interface BatchDownloadFileParams {
  body: {
    fileids: Array<number>
    groupid: number,
    csrfmiddlewaretoken: string
  }
  headers: Headers
}